Separate Arena Experiment submission layers
Keep the typed submission API independent of Hydra and make the renamed CLI own configuration discovery and composition. Remove the test-only main branch and clarify when no co-scheduled policy server is needed.
